requirements.txt
streamlit~=1.37.1
PyYAML~=6.0.1
aiofiles~=24.1.0